backend/requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin

torch==2.5.1
torchaudio==2.5.1

faster-whisper
pyannote.audio==3.3.2
huggingface_hub==0.24.7

pydub

python-dotenv
soundfile

onnxruntime
transformers
tokenizers